INclude DCW index.html file in repo
Because the makefile that updates the DCW website resides in this repo, and the update to the PNG and the ChangeLog is done here, we will place the DCW index.html file here as well so it can easlly be updated at the same time.  For now, only @pwessel can run the make place-dcw command.