Add a clean VERSION file to the tarball and zip file (#65)
In order to implement an update DCW system from the server it will help to have a small file with the version number that we can download upon need to learn if there is a more recent version in the cloud.  I will manually copy over that file to the server for 2.1.0 for now.